refactor: move PasswordCard styles and components to separate directory
