+++ frontend/src/pages/password/passwordCard/PasswordCard.tsx
@@ -1,33 +1,26 @@
-import AddIcon from '@mui/icons-material/Add';
 import ContentCopyIcon from '@mui/icons-material/ContentCopy';
 import Box from '@mui/material/Box';
-import CircularProgress from '@mui/material/CircularProgress';
 import Divider from '@mui/material/Divider';
-import Fab from '@mui/material/Fab';
 import Paper from '@mui/material/Paper';
 import Stack from '@mui/material/Stack';
 import TextField from '@mui/material/TextField';
 import Typography from '@mui/material/Typography';
 import axios from 'axios';
 import classnames from 'classnames/bind';
-import { useEffect, useState } from 'react';
+import { useEffect } from 'react';
 import { useParams } from 'react-router';
 
-import { usePasswordsStore, type Password, type PasswordField } from '@store/passwords';
+import { usePasswordsStore, type Password } from '@store/passwords';
 import { useUserStore } from '@store/user';
 
+import { AddField } from './AddField';
 import * as defaultStyle from './style.scss';
 
 const cx = classnames.bind(defaultStyle);
 
 export function PasswordCard() {
-  const [fieldName, setFieldName] = useState<string>('');
-  const [fieldValue, setFieldValue] = useState<string>('');
-  const [newFieldError, setNewFieldError] = useState<string>('');
-  const [isLoading, setIsLoading] = useState<boolean>(false);
-
   const { passwordId } = useParams();
-  const { currentPassword, setCurrentPassword, addField } = usePasswordsStore();
+  const { currentPassword, setCurrentPassword } = usePasswordsStore();
   const { accessToken } = useUserStore();
 
   useEffect(() => {
@@ -47,45 +40,6 @@ export function PasswordCard() {
   }, [passwordId]);
 
   if (!currentPassword) return null;
-
-  const handleAddField = async () => {
-    if (isLoading) return;
-    if (!fieldName || !fieldValue) {
-      setNewFieldError('Please fill in all fields');
-      return;
-    }
-    setIsLoading(true);
-    try {
-      const response = await axios.post<PasswordField>(
-        `${import.meta.env.VITE_API_HOST}/api/v1/field`,
-        {
-          name: fieldName,
-          value: fieldValue,
-          passwordId: currentPassword.id,
-        },
-        {
-          headers: {
-            Authorization: `Bearer ${accessToken}`,
-          },
-        }
-      );
-      addField({
-        id: response.data.id,
-        name: response.data.name,
-        valueDecrypted: response.data.valueDecrypted,
-        createdAt: response.data.createdAt,
-        isDeleted: false,
-      });
-      setFieldName('');
-      setFieldValue('');
-      setNewFieldError('');
-    } catch (error) {
-      console.error(error);
-      setNewFieldError('Failed to add field');
-    } finally {
-      setIsLoading(false);
-    }
-  };
 
   return (
     <Box sx={{ padding: '1rem' }}>
@@ -119,41 +73,7 @@ export function PasswordCard() {
               />
             ))}
         </Stack>
-        <Stack spacing={2} sx={{ marginTop: '2rem' }} direction="row">
-          <TextField
-            label="New field name"
-            variant="outlined"
-            value={fieldName}
-            fullWidth
-            onChange={(e) => setFieldName(e.target.value)}
-            error={!!newFieldError}
-            helperText={newFieldError}
-          />
-          <TextField
-            label="New field value"
-            variant="outlined"
-            value={fieldValue}
-            fullWidth
-            onChange={(e) => setFieldValue(e.target.value)}
-            error={!!newFieldError}
-            helperText={newFieldError}
-          />
-          <Box sx={{ display: 'flex', alignItems: 'center' }}>
-            {isLoading ? (
-              <CircularProgress size={20} />
-            ) : (
-              <Fab
-                color="primary"
-                aria-label="add"
-                onClick={handleAddField}
-                size="small"
-                disabled={isLoading}
-              >
-                <AddIcon />
-              </Fab>
-            )}
-          </Box>
-        </Stack>
+        <AddField />
         <Divider sx={{ marginTop: '2rem' }}>Deleted Fields</Divider>
         <Stack spacing={2} sx={{ marginTop: '2rem' }}>
           {currentPassword.fields
