+++ frontend/src/pages/password/passwordCard/index.ts
@@ -1,0 +1,3 @@
+import { PasswordCard } from './PasswordCard';
+
+export { PasswordCard };
