+++ frontend/src/pages/password/passwordCard/style.scss
@@ -1,0 +1,39 @@
+.password-card-header {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+
+  img {
+    margin-right: 0.5rem;
+    width: 32px;
+    height: 32px;
+  }
+}
+
+@keyframes clickPop {
+  0% {
+    transform: scale(1);
+  }
+  50% {
+    transform: scale(0.85);
+  }
+  100% {
+    transform: scale(1);
+  }
+}
+
+.copy-icon {
+  cursor: pointer;
+  color: #000;
+  transition: all 0.3s ease;
+
+  &:hover {
+    color: #666;
+    transform: scale(1.1);
+  }
+
+  &:active {
+    animation: clickPop 0.3s ease;
+    color: #4caf50;
+  }
+}
