+++ frontend/src/pages/password/style.scss
@@ -22,18 +22,6 @@
   }
 }
 
-.password-card-header {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-
-  img {
-    margin-right: 0.5rem;
-    width: 32px;
-    height: 32px;
-  }
-}
-
 .passwords-list-autocomplete-option {
   display: flex;
   align-items: center;
@@ -57,33 +45,5 @@
       width: 20px;
       height: 20px;
     }
-  }
-}
-
-@keyframes clickPop {
-  0% {
-    transform: scale(1);
-  }
-  50% {
-    transform: scale(0.85);
-  }
-  100% {
-    transform: scale(1);
-  }
-}
-
-.copy-icon {
-  cursor: pointer;
-  color: #000;
-  transition: all 0.3s ease;
-
-  &:hover {
-    color: #666;
-    transform: scale(1.1);
-  }
-
-  &:active {
-    animation: clickPop 0.3s ease;
-    color: #4caf50;
   }
 }
